requirements.txt
mcp
python-telegram-bot
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv
uvicorn
starlette
requests
